modularized renders on teamsAndResults changes. New custom events and eventListeners were utilized. Additionally a TeamsAndResults Module was created to handle those.
