+++ src/scripts/TruncheonsAndFlagons.js
@@ -1,7 +1,4 @@
-import { RecentGamesList } from "./TeamAndResults/RecentGamesList.js"
-import { JoinTeam } from "./TeamAndResults/TeamStuff/JoinTeam.js";
-import { CreateTeam } from "./TeamAndResults/TeamStuff/CreateTeam.js" 
-import { TeamList } from "./TeamAndResults/TeamStuff/TeamList.js";
+import { TeamsAndResults } from "./TeamsAndResults/TeamsAndResults.js";
 import { Game } from "./GameplayArea/GameModules/Game.js";
 
 export const TruncheonsAndFlagons = () => {
@@ -9,22 +6,8 @@ export const TruncheonsAndFlagons = () => {
     return `
         <h1>Truncheons And Flagons</h1>
         <article class="teamsAndResults containerChunk">
-            <section class="teamFormation teamsAndResultsChunk">
-                <div class="teamFormationChunk teamCreation">
-                    ${CreateTeam()}
-                </div>
-                <div class="teamFormationChunk joinTeam">
-                    ${JoinTeam()}
-                </div>
-                <div class="teamFormationChunk incompleteTeamList">
-                    ${TeamList()}
-                </div>
-            </section>
-            <section class="recentGameResults teamsAndResultsChunk">
-                    ${RecentGamesList()}
-            </section>
+        ${TeamsAndResults()}
         </article>
-
         <article class="gameplayArea containerChunk">
             <section class="gameplayAreaChunk activeScoreBoard">
                 Active Game Scores Go Here
